requirements.txt
streamlit
pandas
vcfpy
reportlab
